requirements.txt
validators
pytorch
torchvision
scikit-image
pillow
matplotlib
shapely